# docs/openapi.yml
openapi: 3.0.3
info:
  title: AI Developer Growth Coach API
  version: v1
  description: |
    v1 API contract for the AI developer growth coach service.

    Source documents:
    - docs/03_api_spec_aligned.md
    - docs/09_contract_appendix.md
    - docs/20_v2_coach_api_contract.md
    - docs/21_v2_db_schema_ddl.md

    The custom extension `x-implementation-status` marks whether an API is
    currently implemented or still planned.
servers:
  - url: http://localhost:8080
    description: Local backend
security:
  - bearerAuth: []
...
paths:
...
  /api/coach/sessions:
    get:
      tags:
        - Coach
      summary: List Coach sessions
      operationId: listCoachSessions
      x-implementation-status: implemented
      responses:
        '200':
          description: Coach sessions sorted by startedAt descending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoachSessionListApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Coach
      summary: Start Coach session
      operationId: startCoachSession
      x-implementation-status: implemented
      responses:
        '201':
          description: Coach session created with pinned snapshot versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoachSessionApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /api/coach/sessions/{sessionId}:
    delete:
      tags:
        - Coach
      summary: Close Coach session
      operationId: closeCoachSession
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/CoachSessionId'
      responses:
        '204':
          description: Coach session closed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /api/coach/sessions/{sessionId}/messages:
    get:
      tags:
        - Coach
      summary: List Coach session messages
      operationId: listCoachSessionMessages
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/CoachSessionId'
      responses:
        '200':
          description: Coach messages sorted by createdAt ascending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoachMessageHistoryListApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'

    post:
      tags:
        - Coach
      summary: Send Coach message
      operationId: sendCoachMessage
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/CoachSessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachMessageRequest'
      responses:
        '200':
          description: Coach reply
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoachMessageApiResponse'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CoachSessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      example: '10001'
...
  responses:
    InvalidInput:
      description: Request value is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication is required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: User does not have permission for the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ResourceNotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with the current state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadGateway:
      description: Upstream LLM or planner call failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    GatewayTimeout:
      description: Upstream LLM call timed out
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ApiMeta:
      type: object
      additionalProperties: true
      example: {}
    ErrorCode:
      type: string
      enum:
        - INVALID_INPUT
        - UNAUTHORIZED
        - FORBIDDEN
        - RESOURCE_NOT_FOUND
        - CONFLICT
        - ANALYSIS_FAILED
        - ROADMAP_GENERATION_FAILED
        - SNAPSHOT_NOT_FOUND
        - SESSION_NOT_FOUND
        - SESSION_CLOSED
        - PROPOSAL_NOT_FOUND
        - PROPOSAL_EXPIRED
        - LLM_TIMEOUT
        - LLM_INVALID_RESPONSE
        - PLANNER_FAILED
        - INTERNAL_SERVER_ERROR
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - details
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          example: 요청 값이 올바르지 않습니다.
        details:
          type: object
          nullable: true
          additionalProperties: true
          example:
            field: currentLevel
...
    CoachRoute:
      type: string
      enum:
        - SIMPLE_GUIDE
        - REPLAN_SUGGEST
        - REPLAN_EXECUTE
        - DISMISS
    ChatSessionStatus:
      type: string
      enum:
        - ACTIVE
        - CLOSED
    CoachMessageRole:
      type: string
      enum:
        - USER
        - COACH
        - SUMMARY
...
    CoachSessionResponseData:
      type: object
      required:
        - sessionId
        - profileVersion
        - roadmapVersion
        - startedAt
      properties:
        sessionId:
          type: string
          example: '10001'
        profileVersion:
          type: integer
          example: 3
        roadmapVersion:
          type: integer
          example: 2
        startedAt:
          type: string
          format: date-time
          example: '2026-05-07T09:00:00Z'
    CoachSessionApiResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/CoachSessionResponseData'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    CoachSessionSummaryData:
      type: object
      required:
        - sessionId
        - profileVersion
        - roadmapVersion
        - status
        - startedAt
      properties:
        sessionId:
          type: string
          example: '10001'
        profileVersion:
          type: integer
          example: 3
        roadmapVersion:
          type: integer
          example: 2
        status:
          $ref: '#/components/schemas/ChatSessionStatus'
        startedAt:
          type: string
          format: date-time
          example: '2026-05-07T09:00:00Z'
        endedAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-05-07T10:00:00Z'
    CoachSessionListApiResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CoachSessionSummaryData'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    CoachMessageRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 4000
          example: 오늘 무엇부터 공부하면 좋을까?
    ReplanProposalSummary:
      type: object
      required:
        - proposalId
        - reason
        - expiresAt
      properties:
        proposalId:
          type: string
          example: '7001'
        reason:
          type: string
          example: 3일 연속 미달성 감지
        expiresAt:
          type: string
          format: date-time
          example: '2026-05-08T09:00:00Z'
    CoachMessageResponseData:
      type: object
      required:
        - messageId
        - responseText
        - route
        - replanProposal
      properties:
        messageId:
          type: string
          example: '9001'
        responseText:
          type: string
          example: 이번 주는 Redis TTL과 캐시 무효화 개념부터 정리하는 것이 좋습니다.
        route:
          $ref: '#/components/schemas/CoachRoute'
        replanProposal:
          allOf:
            - $ref: '#/components/schemas/ReplanProposalSummary'
          nullable: true
    CoachMessageApiResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/CoachMessageResponseData'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    CoachMessageHistoryData:
      type: object
      required:
        - messageId
        - role
        - messageText
        - createdAt
      properties:
        messageId:
          type: string
          example: '9001'
        role:
          $ref: '#/components/schemas/CoachMessageRole'
        messageText:
          type: string
          example: 오늘 무엇부터 공부하면 좋을까?
        route:
          allOf:
            - $ref: '#/components/schemas/CoachRoute'
          nullable: true
        detectedIntent:
          type: string
          nullable: true
          example: CHECK_TODAY_PLAN
        createdAt:
          type: string
          format: date-time
          example: '2026-05-08T09:00:00Z'
    CoachMessageHistoryListApiResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CoachMessageHistoryData'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...